+++ src/modules/RuleEngineValidation.sol
@@ -7,7 +7,7 @@ import "./RuleInternal.sol";
 import "./RuleEngineValidationCommon.sol";
 import "../interfaces/IRuleEngineValidation.sol";
 import "../interfaces/IRuleValidation.sol";
-import "CMTAT/interfaces/draft-IERC1404/draft-IERC1404EnumCode.sol";
+import "CMTAT/interfaces/tokenization/draft-IERC1404.sol";
 
 /**
  * @title Implementation of a ruleEngine defined by the CMTAT
@@ -16,8 +16,7 @@ abstract contract RuleEngineValidation is
     AccessControl,
     RuleInternal,
     RuleEngineValidationCommon,
-    IRuleEngineValidation,
-    IERC1404EnumCode
+    IRuleEngineValidation
 {
     /**
      * @notice Go through all the rule to know if a restriction exists on the transfer
@@ -40,7 +39,24 @@ abstract contract RuleEngineValidation is
             }
         }
 
-        return uint8(REJECTED_CODE_BASE.TRANSFER_OK);
+        return uint8(IERC1404Extend.REJECTED_CODE_BASE.TRANSFER_OK);
+    }
+
+    function detectTransferRestrictionValidationFrom(
+        address spender,
+        address _from,
+        address _to,
+        uint256 _amount
+    ) public view override returns (uint8) {
+        uint256 rulesLength = _rulesValidation.length;
+        for (uint256 i = 0; i < rulesLength; ++i) {
+            uint8 restriction = IRuleValidation(_rulesValidation[i])
+                .detectTransferRestrictionFrom(spender, _from, _to, _amount);
+            if (restriction > 0) {
+                return restriction;
+            }
+        }
+        return uint8(IERC1404Extend.REJECTED_CODE_BASE.TRANSFER_OK);
     }
 
     /**
@@ -50,13 +66,31 @@ abstract contract RuleEngineValidation is
      * @param _amount to transfer
      * @return True if the transfer is valid, false otherwise
      **/
-    function validateTransferValidation(
+    function canTransferValidation(
+        address _from,
+        address _to,
+        uint256 _amount
+    ) public view returns (bool) {
+        return
+            detectTransferRestrictionValidation(_from, _to, _amount) ==
+            uint8(IERC1404Extend.REJECTED_CODE_BASE.TRANSFER_OK);
+    }
+
+    /**
+     * @notice Validate a transfer
+     * @param _from the origin address
+     * @param _to the destination address
+     * @param _amount to transfer
+     * @return True if the transfer is valid, false otherwise
+     **/
+    function canTransferValidationFrom(
+        address spender,
         address _from,
         address _to,
         uint256 _amount
     ) public view override returns (bool) {
         return
-            detectTransferRestrictionValidation(_from, _to, _amount) ==
-            uint8(REJECTED_CODE_BASE.TRANSFER_OK);
+            detectTransferRestrictionValidationFrom(spender, _from, _to, _amount) ==
+            uint8(IERC1404Extend.REJECTED_CODE_BASE.TRANSFER_OK);
     }
 }