+++ src/rules/validation/abstract/RuleValidateTransfer.sol
@@ -2,8 +2,8 @@
 
 pragma solidity ^0.8.20;
 
-import "../../../interfaces/IRuleValidation.sol";
-
+import {IRuleValidation} from "../../../interfaces/IRuleValidation.sol";
+import {IERC1404} from "CMTAT/interfaces/tokenization/draft-IERC1404.sol";
 abstract contract RuleValidateTransfer is IRuleValidation {
     /**
      * @notice Validate a transfer
@@ -12,14 +12,24 @@ abstract contract RuleValidateTransfer is IRuleValidation {
      * @param _amount to transfer
      * @return isValid => true if the transfer is valid, false otherwise
      **/
-    function validateTransfer(
+    function canTransfer(
         address _from,
         address _to,
         uint256 _amount
     ) public view override returns (bool isValid) {
-        // does not work without this keyword "Undeclared identifier"
+        // does not work without `this` keyword => "Undeclared identifier"
         return
             this.detectTransferRestriction(_from, _to, _amount) ==
+            uint8(REJECTED_CODE_BASE.TRANSFER_OK);
+    }
+
+    function canTransferFrom(
+        address spender,
+        address from,
+        address to,
+        uint256 value
+    ) public view virtual override returns (bool) {
+        return this.detectTransferRestrictionFrom(spender, from, to, value)  ==
             uint8(REJECTED_CODE_BASE.TRANSFER_OK);
     }
 }